fixed backwards seeking in dvd player
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@3474 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDInputStreams/DVDInputStream.h
@@ -12,7 +12,7 @@ public:
   virtual bool Open(const char* strFile) = 0;
   virtual void Close() = 0;
   virtual int Read(BYTE* buf, int buf_size) = 0;
-  virtual int Seek(__int64 offset, int whence) = 0;
+  virtual __int64 Seek(__int64 offset, int whence) = 0;
 
   const char* GetFileName();
