+++ xbmc/cores/dvdplayer/DVDInputStreams/DVDInputStreamFile.h
@@ -11,7 +11,7 @@ public:
   virtual bool Open(const char* strFile);
   virtual void Close();
   virtual int Read(BYTE* buf, int buf_size);
-  virtual int Seek(__int64 offset, int whence);
+  virtual __int64 Seek(__int64 offset, int whence);
 
 protected:
   CFile* m_pFile;
